Fixed the adjustable-ratio widget, mostly.
Still needs some tweaking but I'm leaving it here for the night. Moved the DraggableWeightView stuff into MapActivity to play around with it more, tried scaling the touch stuff. Looks like the touch X-Y coordinates and the View height X-Y coordinates are on a different system, which explains why it doesn't look like it's following the finger when it moves.